Actually fix the GCC build, follow up for 12ae813f8e9bebe096d6486b4435686a60e9eb23

+++ src/server/game/AI/SmartScripts/SmartScriptMgr.cpp
@@ -792,7 +792,7 @@ bool SmartAIMgr::CheckUnusedEventParams(SmartScriptHolder const& e)
             if (value != 0)
             {
                 TC_LOG_WARN("sql.sql", "SmartAIMgr: Entry %d SourceType %u Event %u Action %u has unused event_param%u with value %u, it should be 0.",
-                    e.entryOrGuid, e.GetScriptType(), e.event_id, e.GetActionType(), index + 1, value);
+                    e.entryOrGuid, e.GetScriptType(), e.event_id, e.GetActionType(), static_cast<uint32>(index + 1), value);
             }
         }
